git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/java/branches/mrh@999573 13f79535-47bb-0310-9956-ffa450edef68

+++ classlib/modules/pack200/src/main/java/org/apache/harmony/unpack200/Archive.java
@@ -187,8 +187,11 @@ public class Archive {
             }
         }
         if (removePackFile) {
-            File file = new File(inputFileName);
-            boolean deleted = file.delete();
+            boolean deleted = false;
+            if(inputFileName != null) {
+                File file = new File(inputFileName);
+                deleted = file.delete();
+            }
             if (!deleted) {
                 throw new Pack200Exception("Failed to delete the input file.");
             }
@@ -234,7 +237,7 @@ public class Archive {
 
     public void setLogFile(String logFileName, boolean append)
             throws FileNotFoundException {
-        logFile = new FileOutputStream(logFileName, true);
+        logFile = new FileOutputStream(logFileName, append);
     }
 
     public void setDeflateHint(boolean deflateHint) {
